fix LP64 portability problem, from Yamamoto-san in [gfarm-developers:00973]
git-svn-id: https://svn.code.sf.net/p/gfarm/code/gfarm/trunk@2994 4a5e1ece-6a47-475c-b646-731d720b41d0

+++ lib/libgfarm/gfarm/metadb_pgsql.c
@@ -1063,7 +1063,7 @@ gfarm_pgsql_path_info_remove(const char *pathname)
 #define COPY_BINARY(data, buf, residual, msg) { \
 	if (sizeof(data) > residual) \
 		gflog_fatal(msg ": %d bytes needed, but only %d bytes", \
-		    sizeof(data), residual); \
+		    (int)sizeof(data), residual); \
 	memcpy(&(data), buf, sizeof(data)); \
 	buf += sizeof(data); \
 	residual -= sizeof(data); \
